Wolf Whistle
This file generates a function to produce a wolf whistle using Raspberry Pi. It utilises the note_slide parameter.
After the definition, the function is called twice, once in a thread and once in the mail program, so that the two whistles sound together.
If you view the raw file, you can copy and paste it into a Sonic Pi 2 workspace. After copying, delete the first line ```Ruby and the last line ``` which are there for formatting purposes.