Sonic Pi Morse Code
This file uses Sonic Pi 2 in a rather unusual manner. It takes a text string, and plays it in morse code. It illustrates how you can use Ruby constructs in conjunction with standard Sonic Pi commands.
If you view the raw file you can copy and paste it into a Sonic Pi workspace, and then delete the first ```Ruby and the last line ``` which are there to format the file